Moved notifications into its own view
