+++ Drip/Screens/Notifications/Classes/NotificationsViewProtocol.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+protocol NotificationsViewProtocol: class {
+    var presenter: NotificationsPresenterProtocol! { get set }
+    func updateTitle(title: String)
+
+    func reloadTableView()
+}
