+++ Drip/Screens/Notifications/NotificationsScreenBuilder.swift
@@ -1,0 +1,13 @@
+import UIKit
+
+final class NotificationsScreenBuilder: ScreenBuilder {
+
+    func build() -> NotificationsView {
+        let view = NotificationsView()
+        view.presenter = NotificationsPresenter(view: view)
+
+        view.presenter = NotificationsPresenter(view: view)
+
+        return view
+    }
+}
