+++ Drip/Screens/SettingsDetailView/Classes/SettingsDetailPresenterProtocol.swift
@@ -14,7 +14,6 @@ protocol SettingsDetailPresenterProtocol: class {
 
     func numberOfRowsInSection() -> Int
     func coefficientCellDataForRow(row: Int) -> Beverage
-    func notificationTimeStampForRow(row: Int, completion: @escaping (String) -> Void)
     func setCoefficientBool(isEnabled: Bool)
     func attributionTitleForRow(row: Int) -> String
     func getAttributionURLforRow(row: Int) -> URL?
